requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg[binary]==3.2.4
dj-database-url==2.3.0
sqlparse==0.5.3
gunicorn==23.0.0
whitenoise==6.8.2
python-dotenv==1.0.1
asgiref==3.8.1
typing_extensions==4.12.2